ポストエフェクトプロジェクトを追加 DirectX12用ポストエフェクトクラスを追加 アンオーダードアクセスバッファ基底クラスを追加(DirectX12用クラスの追加) ポストエフェクト(ブルーム)の実装 ブラーやブルーム用のシェーダーを追加 一部コードのを改善

+++ Sources/Engine46/CShaderManager.cpp
@@ -73,7 +73,7 @@ namespace Engine46 {
 
         if (!ofs.is_open()) return false;
 
-        int packageSize = m_pMapShaderPackage.size();
+        int packageSize = (UINT)m_pMapShaderPackage.size();
         ofs.write((char*)&packageSize, sizeof(int));
 
         for (const auto& pSp : m_pMapShaderPackage) {
